refactor(popup): file分割
