+++ packages/wiz-ui-react/src/components/base/popup/types/placement.ts
@@ -1,0 +1,11 @@
+export type PlacementOption = {
+  anchor: DOMRect;
+  content?: DOMRect;
+  gap?: string;
+};
+
+export type PlacementStyle = {
+  top?: number;
+  left?: number;
+  transform?: string;
+};
