+++ packages/wiz-ui-react/src/components/base/popup/utils/index.ts
@@ -1,1 +1,2 @@
+export * from "./wrap";
 export * from "./placement";
