+++ packages/wiz-ui-react/src/components/base/popup/utils/placement/index.ts
@@ -1,0 +1,2 @@
+export * from "./on-child";
+export * from "./on-portal";
